Network.AWS.CodeCommit.CreateBranch

Description

Creates a new branch in a repository and points the branch to a commit.

Creating a Request

createBranch

Arguments

:: Text
cbRepositoryName

-> Text
cbBranchName

-> Text
cbCommitId

-> CreateBranch

Creates a value of CreateBranch with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cbRepositoryName - The name of the repository in which you want to create the new branch.

cbBranchName - The name of the new branch to create.

cbCommitId - The ID of the commit to point the new branch to.

data CreateBranch

Represents the input of a create branch operation.

See: createBranch smart constructor.

Request Lenses

cbRepositoryName :: Lens' CreateBranch Text

The name of the repository in which you want to create the new branch.

cbBranchName :: Lens' CreateBranch Text

The name of the new branch to create.

cbCommitId :: Lens' CreateBranch Text

The ID of the commit to point the new branch to.

Destructuring the Response

createBranchResponse :: CreateBranchResponse

Creates a value of CreateBranchResponse with the minimum fields required to make a request.

data CreateBranchResponse

See: createBranchResponse smart constructor.