Network.AWS.CodeCommit.GetBranch

Description

Returns information about a repository branch, including its name and the last commit ID.

Creating a Request

getBranch :: GetBranch

Creates a value of GetBranch with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gbBranchName - The name of the branch for which you want to retrieve information.

gbRepositoryName - The name of the repository that contains the branch for which you want to retrieve information.

data GetBranch

Represents the input of a get branch operation.

See: getBranch smart constructor.

Request Lenses

gbBranchName :: Lens' GetBranch (Maybe Text)

The name of the branch for which you want to retrieve information.

gbRepositoryName :: Lens' GetBranch (Maybe Text)

The name of the repository that contains the branch for which you want to retrieve information.

Destructuring the Response

getBranchResponse

Arguments

:: Int
grsResponseStatus

-> GetBranchResponse

Creates a value of GetBranchResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

grsBranch - The name of the branch.

grsResponseStatus - -- | The response status code.

data GetBranchResponse

Represents the output of a get branch operation.

See: getBranchResponse smart constructor.

Response Lenses

grsBranch :: Lens' GetBranchResponse (Maybe BranchInfo)

The name of the branch.

grsResponseStatus :: Lens' GetBranchResponse Int

- | The response status code.