Network.AWS.CodeCommit.GetDifferences

Description

Returns information about the differences in a valid commit specifier (such as a branch, tag, HEAD, commit ID or other fully qualified reference). Results can be limited to a specified path.

Creating a Request

getDifferences

Arguments

:: Text
gdRepositoryName

-> Text
gdAfterCommitSpecifier

-> GetDifferences

Creates a value of GetDifferences with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gdAfterPath - The file path in which to check differences. Limits the results to this path. Can also be used to specify the changed name of a directory or folder, if it has changed. If not specified, differences will be shown for all paths.

gdNextToken - An enumeration token that when provided in a request, returns the next batch of the results.

gdBeforeCommitSpecifier - The branch, tag, HEAD, or other fully qualified reference used to identify a commit. For example, the full commit ID. Optional. If not specified, all changes prior to the afterCommitSpecifier value will be shown. If you do not use beforeCommitSpecifier in your request, consider limiting the results with maxResults .

gdBeforePath - The file path in which to check for differences. Limits the results to this path. Can also be used to specify the previous name of a directory or folder. If beforePath and afterPath are not specified, differences will be shown for all paths.

gdMaxResults - A non-negative integer used to limit the number of returned results.

gdRepositoryName - The name of the repository where you want to get differences.

gdAfterCommitSpecifier - The branch, tag, HEAD, or other fully qualified reference used to identify a commit.

data GetDifferences

See: getDifferences smart constructor.

Request Lenses

gdAfterPath :: Lens' GetDifferences (Maybe Text)

The file path in which to check differences. Limits the results to this path. Can also be used to specify the changed name of a directory or folder, if it has changed. If not specified, differences will be shown for all paths.

gdNextToken :: Lens' GetDifferences (Maybe Text)

An enumeration token that when provided in a request, returns the next batch of the results.

gdBeforeCommitSpecifier :: Lens' GetDifferences (Maybe Text)

The branch, tag, HEAD, or other fully qualified reference used to identify a commit. For example, the full commit ID. Optional. If not specified, all changes prior to the afterCommitSpecifier value will be shown. If you do not use beforeCommitSpecifier in your request, consider limiting the results with maxResults .

gdBeforePath :: Lens' GetDifferences (Maybe Text)

The file path in which to check for differences. Limits the results to this path. Can also be used to specify the previous name of a directory or folder. If beforePath and afterPath are not specified, differences will be shown for all paths.

gdMaxResults :: Lens' GetDifferences (Maybe Int)

A non-negative integer used to limit the number of returned results.

gdRepositoryName :: Lens' GetDifferences Text

The name of the repository where you want to get differences.

gdAfterCommitSpecifier :: Lens' GetDifferences Text

The branch, tag, HEAD, or other fully qualified reference used to identify a commit.

Destructuring the Response

getDifferencesResponse

Arguments

:: Int
gdrsResponseStatus

-> GetDifferencesResponse

Creates a value of GetDifferencesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gdrsNextToken - An enumeration token that can be used in a request to return the next batch of the results.

gdrsDifferences - A differences data type object that contains information about the differences, including whether the difference is added, modified, or deleted (A, D, M).

gdrsResponseStatus - -- | The response status code.

data GetDifferencesResponse

See: getDifferencesResponse smart constructor.

Response Lenses

gdrsNextToken :: Lens' GetDifferencesResponse (Maybe Text)

An enumeration token that can be used in a request to return the next batch of the results.

gdrsDifferences :: Lens' GetDifferencesResponse [Difference]

A differences data type object that contains information about the differences, including whether the difference is added, modified, or deleted (A, D, M).

gdrsResponseStatus :: Lens' GetDifferencesResponse Int

- | The response status code.