Network.AWS.CodeCommit.GetRepositoryTriggers

Description

Gets information about triggers configured for a repository.

Creating a Request

getRepositoryTriggers

Arguments

:: Text
grtRepositoryName

-> GetRepositoryTriggers

Creates a value of GetRepositoryTriggers with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

grtRepositoryName - The name of the repository for which the trigger is configured.

data GetRepositoryTriggers

Represents the input of a get repository triggers operation.

See: getRepositoryTriggers smart constructor.

Request Lenses

grtRepositoryName :: Lens' GetRepositoryTriggers Text

The name of the repository for which the trigger is configured.

Destructuring the Response

getRepositoryTriggersResponse

Arguments

:: Int
grtrsResponseStatus

-> GetRepositoryTriggersResponse

Creates a value of GetRepositoryTriggersResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

grtrsConfigurationId - The system-generated unique ID for the trigger.

grtrsTriggers - The JSON block of configuration information for each trigger.

grtrsResponseStatus - -- | The response status code.

data GetRepositoryTriggersResponse

Represents the output of a get repository triggers operation.

See: getRepositoryTriggersResponse smart constructor.

Response Lenses

grtrsConfigurationId :: Lens' GetRepositoryTriggersResponse (Maybe Text)

The system-generated unique ID for the trigger.

grtrsTriggers :: Lens' GetRepositoryTriggersResponse [RepositoryTrigger]

The JSON block of configuration information for each trigger.

grtrsResponseStatus :: Lens' GetRepositoryTriggersResponse Int

- | The response status code.