Network.AWS.CodeCommit.TestRepositoryTriggers

Description

Tests the functionality of repository triggers by sending information to the trigger target. If real data is available in the repository, the test will send data from the last commit. If no data is available, sample data will be generated.

Creating a Request

testRepositoryTriggers

Arguments

:: Text
trtRepositoryName

-> TestRepositoryTriggers

Creates a value of TestRepositoryTriggers with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

trtRepositoryName - The name of the repository in which to test the triggers.

trtTriggers - The list of triggers to test.

data TestRepositoryTriggers

Represents the input of a test repository triggers operation.

See: testRepositoryTriggers smart constructor.

Request Lenses

trtRepositoryName :: Lens' TestRepositoryTriggers Text

The name of the repository in which to test the triggers.

trtTriggers :: Lens' TestRepositoryTriggers [RepositoryTrigger]

The list of triggers to test.

Destructuring the Response

testRepositoryTriggersResponse

Arguments

:: Int
trtrsResponseStatus

-> TestRepositoryTriggersResponse

Creates a value of TestRepositoryTriggersResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

trtrsFailedExecutions - The list of triggers that were not able to be tested. This list provides the names of the triggers that could not be tested, separated by commas.

trtrsSuccessfulExecutions - The list of triggers that were successfully tested. This list provides the names of the triggers that were successfully tested, separated by commas.

trtrsResponseStatus - -- | The response status code.

data TestRepositoryTriggersResponse

Represents the output of a test repository triggers operation.

See: testRepositoryTriggersResponse smart constructor.

Response Lenses

trtrsFailedExecutions :: Lens' TestRepositoryTriggersResponse [RepositoryTriggerExecutionFailure]

The list of triggers that were not able to be tested. This list provides the names of the triggers that could not be tested, separated by commas.

trtrsSuccessfulExecutions :: Lens' TestRepositoryTriggersResponse [Text]

The list of triggers that were successfully tested. This list provides the names of the triggers that were successfully tested, separated by commas.

trtrsResponseStatus :: Lens' TestRepositoryTriggersResponse Int

- | The response status code.