Network.AWS.CodeCommit.Types

Description

Service Configuration

codeCommit :: Service

API version 2015-04-13 of the Amazon CodeCommit SDK configuration.

Errors

_InvalidRepositoryTriggerRegionException :: AsError a => Getting (First ServiceError) a ServiceError

The region for the trigger target does not match the region for the repository. Triggers must be created in the same region as the target for the trigger.

_InvalidContinuationTokenException :: AsError a => Getting (First ServiceError) a ServiceError

The specified continuation token is not valid.

_EncryptionKeyNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

No encryption key was found.

_InvalidRepositoryTriggerBranchNameException :: AsError a => Getting (First ServiceError) a ServiceError

One or more branch names specified for the trigger is not valid.

_InvalidRepositoryTriggerCustomDataException :: AsError a => Getting (First ServiceError) a ServiceError

The custom data provided for the trigger is not valid.

_BlobIdDoesNotExistException :: AsError a => Getting (First ServiceError) a ServiceError

The specified blob does not exist.

_MaximumRepositoryNamesExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The maximum number of allowed repository names was exceeded. Currently, this number is 25.

_InvalidRepositoryDescriptionException :: AsError a => Getting (First ServiceError) a ServiceError

The specified repository description is not valid.

_RepositoryNameExistsException :: AsError a => Getting (First ServiceError) a ServiceError

The specified repository name already exists.

_MaximumRepositoryTriggersExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The number of triggers allowed for the repository was exceeded.

_InvalidBranchNameException :: AsError a => Getting (First ServiceError) a ServiceError

The specified branch name is not valid.

_BranchNameRequiredException :: AsError a => Getting (First ServiceError) a ServiceError

A branch name is required but was not specified.

_RepositoryTriggersListRequiredException :: AsError a => Getting (First ServiceError) a ServiceError

The list of triggers for the repository is required but was not specified.

_EncryptionKeyUnavailableException :: AsError a => Getting (First ServiceError) a ServiceError

The encryption key is not available.

_InvalidRepositoryTriggerDestinationARNException :: AsError a => Getting (First ServiceError) a ServiceError

The Amazon Resource Name (ARN) for the trigger is not valid for the specified destination. The most common reason for this error is that the ARN does not meet the requirements for the service type.

_BlobIdRequiredException :: AsError a => Getting (First ServiceError) a ServiceError

A blob ID is required but was not specified.

_RepositoryNamesRequiredException :: AsError a => Getting (First ServiceError) a ServiceError

A repository names object is required but was not specified.

_InvalidBlobIdException :: AsError a => Getting (First ServiceError) a ServiceError

The specified blob is not valid.

_InvalidOrderException :: AsError a => Getting (First ServiceError) a ServiceError

The specified sort order is not valid.

_BranchDoesNotExistException :: AsError a => Getting (First ServiceError) a ServiceError

The specified branch does not exist.

_InvalidPathException :: AsError a => Getting (First ServiceError) a ServiceError

The specified path is not valid.

_RepositoryTriggerNameRequiredException :: AsError a => Getting (First ServiceError) a ServiceError

A name for the trigger is required but was not specified.

_RepositoryDoesNotExistException :: AsError a => Getting (First ServiceError) a ServiceError

The specified repository does not exist.

_MaximumBranchesExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The number of branches for the trigger was exceeded.

_PathDoesNotExistException :: AsError a => Getting (First ServiceError) a ServiceError

The specified path does not exist.

_EncryptionIntegrityChecksFailedException :: AsError a => Getting (First ServiceError) a ServiceError

An encryption integrity check failed.

_RepositoryTriggerEventsListRequiredException :: AsError a => Getting (First ServiceError) a ServiceError

At least one event for the trigger is required but was not specified.

_EncryptionKeyAccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

An encryption key could not be accessed.

_BranchNameExistsException :: AsError a => Getting (First ServiceError) a ServiceError

The specified branch name already exists.

_InvalidCommitException :: AsError a => Getting (First ServiceError) a ServiceError

The specified commit is not valid.

_InvalidSortByException :: AsError a => Getting (First ServiceError) a ServiceError

The specified sort by value is not valid.

_EncryptionKeyDisabledException :: AsError a => Getting (First ServiceError) a ServiceError

The encryption key is disabled.

_CommitRequiredException :: AsError a => Getting (First ServiceError) a ServiceError

A commit was not specified.

_CommitIdRequiredException :: AsError a => Getting (First ServiceError) a ServiceError

A commit ID was not specified.

_InvalidCommitIdException :: AsError a => Getting (First ServiceError) a ServiceError

The specified commit ID is not valid.

_RepositoryTriggerDestinationARNRequiredException :: AsError a => Getting (First ServiceError) a ServiceError

A destination ARN for the target service for the trigger is required but was not specified.

_CommitDoesNotExistException :: AsError a => Getting (First ServiceError) a ServiceError

The specified commit does not exist or no commit was specified, and the specified repository has no default branch.

_RepositoryTriggerBranchNameListRequiredException :: AsError a => Getting (First ServiceError) a ServiceError

At least one branch name is required but was not specified in the trigger configuration.

_InvalidMaxResultsException :: AsError a => Getting (First ServiceError) a ServiceError

The specified number of maximum results is not valid.

_FileTooLargeException :: AsError a => Getting (First ServiceError) a ServiceError

The specified file exceeds the file size limit for AWS CodeCommit. For more information about limits in AWS CodeCommit, see AWS CodeCommit User Guide .

_CommitIdDoesNotExistException :: AsError a => Getting (First ServiceError) a ServiceError

The specified commit ID does not exist.

_InvalidRepositoryTriggerNameException :: AsError a => Getting (First ServiceError) a ServiceError

The name of the trigger is not valid.

_RepositoryNameRequiredException :: AsError a => Getting (First ServiceError) a ServiceError

A repository name is required but was not specified.

_RepositoryLimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

A repository resource limit was exceeded.

_InvalidRepositoryTriggerEventsException :: AsError a => Getting (First ServiceError) a ServiceError

One or more events specified for the trigger is not valid. Check to make sure that all events specified match the requirements for allowed events.

_InvalidRepositoryNameException :: AsError a => Getting (First ServiceError) a ServiceError

At least one specified repository name is not valid.

ChangeTypeEnum

data ChangeTypeEnum

Constructors

A

D

M

OrderEnum

data OrderEnum

Constructors

Ascending

Descending

RepositoryTriggerEventEnum

data RepositoryTriggerEventEnum

Constructors

All

CreateReference

DeleteReference

UpdateReference

SortByEnum

data SortByEnum

Constructors

LastModifiedDate

RepositoryName

BlobMetadata

data BlobMetadata

Returns information about a specific Git blob object.

See: blobMetadata smart constructor.

blobMetadata :: BlobMetadata

Creates a value of BlobMetadata with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

bmPath - The path to the blob and any associated file name, if any.

bmMode - The file mode permissions of the blob. File mode permission codes include: * 100644 indicates readwrite * 100755 indicates readwrite/execute * 160000 indicates a submodule * 120000 indicates a symlink

bmBlobId - The full ID of the blob.

bmPath :: Lens' BlobMetadata (Maybe Text)

The path to the blob and any associated file name, if any.

bmMode :: Lens' BlobMetadata (Maybe Text)

The file mode permissions of the blob. File mode permission codes include: * 100644 indicates readwrite * 100755 indicates readwrite/execute * 160000 indicates a submodule * 120000 indicates a symlink

bmBlobId :: Lens' BlobMetadata (Maybe Text)

The full ID of the blob.

BranchInfo

data BranchInfo

Returns information about a branch.

See: branchInfo smart constructor.

branchInfo :: BranchInfo

Creates a value of BranchInfo with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

biCommitId - The ID of the last commit made to the branch.

biBranchName - The name of the branch.

biCommitId :: Lens' BranchInfo (Maybe Text)

The ID of the last commit made to the branch.

biBranchName :: Lens' BranchInfo (Maybe Text)

The name of the branch.

Commit

data Commit

Returns information about a specific commit.

See: commit smart constructor.

commit :: Commit

Creates a value of Commit with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cCommitter - Information about the person who committed the specified commit, also known as the committer. Information includes the date in timestamp format with GMT offset, the name of the committer, and the email address for the committer, as configured in Git. For more information about the difference between an author and a committer in Git, see Viewing the Commit History in Pro Git by Scott Chacon and Ben Straub.

cTreeId - Tree information for the specified commit.

cAdditionalData - Any additional data associated with the specified commit.

cParents - The parent list for the specified commit.

cAuthor - Information about the author of the specified commit. Information includes the date in timestamp format with GMT offset, the name of the author, and the email address for the author, as configured in Git.

cMessage - The commit message associated with the specified commit.

cCommitter :: Lens' Commit (Maybe UserInfo)

Information about the person who committed the specified commit, also known as the committer. Information includes the date in timestamp format with GMT offset, the name of the committer, and the email address for the committer, as configured in Git. For more information about the difference between an author and a committer in Git, see Viewing the Commit History in Pro Git by Scott Chacon and Ben Straub.

cTreeId :: Lens' Commit (Maybe Text)

Tree information for the specified commit.

cAdditionalData :: Lens' Commit (Maybe Text)

Any additional data associated with the specified commit.

cParents :: Lens' Commit [Text]

The parent list for the specified commit.

cAuthor :: Lens' Commit (Maybe UserInfo)

Information about the author of the specified commit. Information includes the date in timestamp format with GMT offset, the name of the author, and the email address for the author, as configured in Git.

cMessage :: Lens' Commit (Maybe Text)

The commit message associated with the specified commit.

Difference

data Difference

Returns information about a set of differences for a commit specifier.

See: difference smart constructor.

difference :: Difference

Creates a value of Difference with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dAfterBlob - Information about an afterBlob data type object, including the ID, the file mode permission code, and the path.

dBeforeBlob - Information about a beforeBlob data type object, including the ID, the file mode permission code, and the path.

dChangeType - Whether the change type of the difference is an addition (A), deletion (D), or modification (M).

dAfterBlob :: Lens' Difference (Maybe BlobMetadata)

Information about an afterBlob data type object, including the ID, the file mode permission code, and the path.

dBeforeBlob :: Lens' Difference (Maybe BlobMetadata)

Information about a beforeBlob data type object, including the ID, the file mode permission code, and the path.

dChangeType :: Lens' Difference (Maybe ChangeTypeEnum)

Whether the change type of the difference is an addition (A), deletion (D), or modification (M).

RepositoryMetadata

data RepositoryMetadata

Information about a repository.

See: repositoryMetadata smart constructor.

repositoryMetadata :: RepositoryMetadata

Creates a value of RepositoryMetadata with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rmRepositoryDescription - A comment or description about the repository.

rmLastModifiedDate - The date and time the repository was last modified, in timestamp format.

rmARN - The Amazon Resource Name (ARN) of the repository.

rmCloneURLHTTP - The URL to use for cloning the repository over HTTPS.

rmAccountId - The ID of the AWS account associated with the repository.

rmDefaultBranch - The repository's default branch name.

rmRepositoryId - The ID of the repository.

rmRepositoryName - The repository's name.

rmCreationDate - The date and time the repository was created, in timestamp format.

rmCloneURLSSH - The URL to use for cloning the repository over SSH.

rmRepositoryDescription :: Lens' RepositoryMetadata (Maybe Text)

A comment or description about the repository.

rmLastModifiedDate :: Lens' RepositoryMetadata (Maybe UTCTime)

The date and time the repository was last modified, in timestamp format.

rmARN :: Lens' RepositoryMetadata (Maybe Text)

The Amazon Resource Name (ARN) of the repository.

rmCloneURLHTTP :: Lens' RepositoryMetadata (Maybe Text)

The URL to use for cloning the repository over HTTPS.

rmAccountId :: Lens' RepositoryMetadata (Maybe Text)

The ID of the AWS account associated with the repository.

rmDefaultBranch :: Lens' RepositoryMetadata (Maybe Text)

The repository's default branch name.

rmRepositoryId :: Lens' RepositoryMetadata (Maybe Text)

The ID of the repository.

rmRepositoryName :: Lens' RepositoryMetadata (Maybe Text)

The repository's name.

rmCreationDate :: Lens' RepositoryMetadata (Maybe UTCTime)

The date and time the repository was created, in timestamp format.

rmCloneURLSSH :: Lens' RepositoryMetadata (Maybe Text)

The URL to use for cloning the repository over SSH.

RepositoryNameIdPair

data RepositoryNameIdPair

Information about a repository name and ID.

See: repositoryNameIdPair smart constructor.

repositoryNameIdPair :: RepositoryNameIdPair

Creates a value of RepositoryNameIdPair with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rnipRepositoryId - The ID associated with the repository.

rnipRepositoryName - The name associated with the repository.

rnipRepositoryId :: Lens' RepositoryNameIdPair (Maybe Text)

The ID associated with the repository.

rnipRepositoryName :: Lens' RepositoryNameIdPair (Maybe Text)

The name associated with the repository.

RepositoryTrigger

data RepositoryTrigger

Information about a trigger for a repository.

See: repositoryTrigger smart constructor.

repositoryTrigger

Arguments

:: Text
rtName

-> Text
rtDestinationARN

-> RepositoryTrigger

Creates a value of RepositoryTrigger with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rtBranches - The branches that will be included in the trigger configuration. If no branches are specified, the trigger will apply to all branches.

rtCustomData - Any custom data associated with the trigger that will be included in the information sent to the target of the trigger.

rtName - The name of the trigger.

rtDestinationARN - The ARN of the resource that is the target for a trigger. For example, the ARN of a topic in Amazon Simple Notification Service (SNS).

rtEvents - The repository events that will cause the trigger to run actions in another service, such as sending a notification through Amazon Simple Notification Service (SNS).

rtBranches :: Lens' RepositoryTrigger [Text]

The branches that will be included in the trigger configuration. If no branches are specified, the trigger will apply to all branches.

rtCustomData :: Lens' RepositoryTrigger (Maybe Text)

Any custom data associated with the trigger that will be included in the information sent to the target of the trigger.

rtName :: Lens' RepositoryTrigger Text

The name of the trigger.

rtDestinationARN :: Lens' RepositoryTrigger Text

The ARN of the resource that is the target for a trigger. For example, the ARN of a topic in Amazon Simple Notification Service (SNS).

rtEvents :: Lens' RepositoryTrigger [RepositoryTriggerEventEnum]

The repository events that will cause the trigger to run actions in another service, such as sending a notification through Amazon Simple Notification Service (SNS).

RepositoryTriggerExecutionFailure

data RepositoryTriggerExecutionFailure

A trigger failed to run.

See: repositoryTriggerExecutionFailure smart constructor.

repositoryTriggerExecutionFailure :: RepositoryTriggerExecutionFailure

Creates a value of RepositoryTriggerExecutionFailure with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rtefFailureMessage - Additional message information about the trigger that did not run.

rtefTrigger - The name of the trigger that did not run.

rtefFailureMessage :: Lens' RepositoryTriggerExecutionFailure (Maybe Text)

Additional message information about the trigger that did not run.

rtefTrigger :: Lens' RepositoryTriggerExecutionFailure (Maybe Text)

The name of the trigger that did not run.

UserInfo

data UserInfo

Information about the user who made a specified commit.

See: userInfo smart constructor.

userInfo :: UserInfo

Creates a value of UserInfo with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uiEmail - The email address associated with the user who made the commit, if any.

uiDate - The date when the specified commit was pushed to the repository.

uiName - The name of the user who made the specified commit.

uiEmail :: Lens' UserInfo (Maybe Text)

The email address associated with the user who made the commit, if any.

uiDate :: Lens' UserInfo (Maybe Text)

The date when the specified commit was pushed to the repository.

uiName :: Lens' UserInfo (Maybe Text)

The name of the user who made the specified commit.