Network.AWS.CodeCommit.UpdateDefaultBranch

Description

Sets or changes the default branch name for the specified repository.

Creating a Request

updateDefaultBranch

Arguments

:: Text
udbRepositoryName

-> Text
udbDefaultBranchName

-> UpdateDefaultBranch

Creates a value of UpdateDefaultBranch with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

udbRepositoryName - The name of the repository to set or change the default branch for.

udbDefaultBranchName - The name of the branch to set as the default.

data UpdateDefaultBranch

Represents the input of an update default branch operation.

See: updateDefaultBranch smart constructor.

Request Lenses

udbRepositoryName :: Lens' UpdateDefaultBranch Text

The name of the repository to set or change the default branch for.

udbDefaultBranchName :: Lens' UpdateDefaultBranch Text

The name of the branch to set as the default.

Destructuring the Response

updateDefaultBranchResponse :: UpdateDefaultBranchResponse

Creates a value of UpdateDefaultBranchResponse with the minimum fields required to make a request.

data UpdateDefaultBranchResponse

See: updateDefaultBranchResponse smart constructor.