Network.AWS.CodeCommit.UpdateRepositoryDescription

Description

Sets or changes the comment or description for a repository.

Creating a Request

updateRepositoryDescription

Arguments

:: Text
urdRepositoryName

-> UpdateRepositoryDescription

Creates a value of UpdateRepositoryDescription with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

urdRepositoryDescription - The new comment or description for the specified repository. Repository descriptions are limited to 1,000 characters.

urdRepositoryName - The name of the repository to set or change the comment or description for.

data UpdateRepositoryDescription

Represents the input of an update repository description operation.

See: updateRepositoryDescription smart constructor.

Request Lenses

urdRepositoryDescription :: Lens' UpdateRepositoryDescription (Maybe Text)

The new comment or description for the specified repository. Repository descriptions are limited to 1,000 characters.

urdRepositoryName :: Lens' UpdateRepositoryDescription Text

The name of the repository to set or change the comment or description for.

Destructuring the Response

updateRepositoryDescriptionResponse :: UpdateRepositoryDescriptionResponse

Creates a value of UpdateRepositoryDescriptionResponse with the minimum fields required to make a request.

data UpdateRepositoryDescriptionResponse

See: updateRepositoryDescriptionResponse smart constructor.